Create proper project structure
