+++ integ_tester/pyproject.toml
@@ -8,6 +8,7 @@ authors = [
 ]
 requires-python = ">=3.14"
 dependencies = [
+    "click>=8.3.1",
     "docker>=7.1.0",
     "pyre-extensions>=0.0.32",
 ]
