+++ integ_tester/src/cluster_manager/drivers/base.py
@@ -1,0 +1,4 @@
+import abc
+
+class BaseDriver(abc.ABC):
+    pass
