+++ integ_tester/src/cluster_manager/local_net.py
@@ -1,0 +1,2 @@
+class LocalNetwork:
+    pass
